BIC-154 # improve storage detection and refactor
